refactor: abstract server creation
