+++ server/index.js
@@ -3,11 +3,7 @@ const createApp = require('./app');
 
 const port = process.env.PORT;
 
-const config = {
-  connectionString: process.env.PG_URI,
-};
-
-createApp(config).then((app) => {
+createApp().then((app) => {
   app.listen(port, () => {
     console.log(`Server running at http://localhost:${port}`);
   });
